+++ interpreter/interpreter.go
@@ -1,4 +1,4 @@
-package main
+package interpreter
 
 import (
 	"fmt"
